unified_api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
openai==1.3.5
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
pydantic==2.3.0
pytest==7.4.0 